backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.0
python-multipart==0.0.6
openai==1.6.1
python-dotenv==1.0.0
fastapi-cors==0.0.6
pymongo==4.6.1